backend/foodgram/requirements.txt
Django==2.2.16
django-extensions==3.1.5
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
django-extra-fields==2.0.2
drf-extra-fields==3.3.0
djoser==2.0.1
Pillow==8.1.0
django-cors-headers==3.3.0
six==1.16.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
asgiref==3.2.10
python-dotenv==0.21.0